Remove obsolete volume-time sensitivity models and related files
- Deleted the scaler_HLV_O3.json and weights_HLV_O3.hdf5 files as they are no longer needed.
- Removed the PopModelsVolumeTimeSensitivity and related classes from _popmodelvt.py.
- Eliminated the SemiAnalyticalRealInjectionVolumeTimeSensitivity and SyntheticInjectionVolumeTimeSensitivity classes from their respective files.
- Cleaned up utility functions and the main _vt.py file to reflect the removal of the above classes.
