+++ pyproject.toml
@@ -8,9 +8,6 @@ include-package-data = true
 [tool.setuptools.packages]
 find = { where = ["src/"] }
 
-[tool.setuptools.package-data]
-"gwkokab.vts._pdet" = ["scaler_HLV_O3.json", "weights_HLV_O3.hdf5"]
-
 [project]
 authors = [
     { name = "Meesum Qazalbash", email = "meesumqazalbash@gmail.com" },
