--- src/gwkokab/vts/_pdet/scaler_HLV_O3.json
@@ -1,16 +1,0 @@
-{
-    "mean": [
-        2.6838227328631987, 1.876082848311286, 15.20160842137149,
-        40.1588654050524, 0.19986034900701877, 0.5624349878182003,
-        2.7862424839367432, 3.164953259203175, 0.0007726398631360905,
-        0.5938959595459761, 0.6370285573050727, -0.0007057577971944004,
-        0.018192425289656134, 0.0072821885848800075, 0.3856012936665666
-    ],
-    "scale": [
-        3.401053666147001, 4.002281337234727, 26.399019281570432,
-        62.13319643100984, 0.06809483029847405, 0.30750006778877237,
-        4.060882873475171, 1.8357926470496169, 0.5987647036943281,
-        0.28904172440182047, 0.30722727758516805, 0.7069692489715201,
-        0.23151894359482683, 0.1908847085857963, 0.2721996598886559
-    ]
-}
